Refactoring | Updated file structure section in README.md
