+++ src/components/App Bar/Navigation Buttons/SaveQuestionnaireButton.js
@@ -1,7 +1,7 @@
 import Button from '@material-ui/core/Button'
 import React from 'react'
 import SaveIcon from '@material-ui/icons/Save'
-import { SafeQuestionnaireDialog } from '../../Home Page/SaveQuestionnaireDialog'
+import { SafeQuestionnaireDialog } from '../../Editor/SaveQuestionnaireDialog'
 
 export function SaveQuestionnaireButton() {
     const [saveQuestionnaireOpen, setSafeQuestionnaireOpen] = React.useState(false)
