+++ src/components/Editor/EditingFeature.js
@@ -2,13 +2,13 @@ import React, {createElement, useEffect} from "react";
 import {Box, Button, Dialog, Divider, Grid, Paper, Typography,} from "@material-ui/core";
 import {CLEAN_SUPER_QUESTION, PROPERTIES_BY_QUESTION_TYPE,} from "../../utils";
 import {useDispatch, useSelector} from "react-redux";
-import {setQuestion,} from "../../features/question/questionSlice";
+import {setQuestion,} from "../../features/State Management/questionSlice";
 import TypeProperty from "../properties/TypeProperty";
 import store from "../../store";
 import {postprocessQuestion} from "../properties/postprocessor";
-import {UPDATE_QUESTION} from "../../features/questions/questionsSlice";
-import {SET_UTILITIES} from "../../features/utilities/utilitiesSlice";
-import {useStyles} from "../Home Page/SaveQuestionnaireDialog"
+import {UPDATE_QUESTION} from "../../features/State Management/questionsSlice";
+import {SET_UTILITIES} from "../../features/State Management/utilitiesSlice";
+import {useStyles} from "./SaveQuestionnaireDialog"
 
 
 const EditDialog2 = ({question, index, open, setOpen}) => {
