+++ src/components/properties/TextArrayTemplate.js
@@ -5,7 +5,7 @@ import {
     selectProperty,
     setTextArrayElement,
     setTextArrayField,
-} from '../../features/question/questionSlice'
+} from '../../features/State Management/questionSlice'
 import {
     Button,
     FilledInput,
