+++ src/components/properties/TextArrayTemplate/TextArrayTemplate.js
@@ -1,7 +1,7 @@
 import React, {useEffect} from "react";
 import {useDispatch, useSelector} from "react-redux";
 import {Box, Grid, makeStyles, Paper, Typography} from "@material-ui/core";
-import {selectProperty, setProperty,} from "../../../features/question/questionSlice";
+import {selectProperty, setProperty,} from "../../../features/State Management/questionSlice";
 import EachOption from "./EachOption";
 import AddOptionButton from "./AddOptionButton";
